requirements.txt
# Core RAG pipeline
llama-index-core==0.14.20
llama-index-vector-stores-chroma==0.5.5
llama-index-embeddings-openai==0.6.0
llama-index-llms-openai==0.7.5
llama-index-readers-file==0.6.0

# Vector store
chromadb==1.5.7

# PDF parsing
pypdf==6.10.0

# OpenAI
openai==2.31.0

# Evaluation
ragas==0.4.3
datasets==4.8.4

# UI
streamlit==1.56.0

# Utilities
python-dotenv==1.2.2
pandas==2.2.3
tqdm==4.67.3
